main func ready for test

+++ headers.h
@@ -18,4 +18,4 @@ complex_num* div_copmplex( complex_num* num1, complex_num* num2);
 
 complex_num* pow_copmplex( complex_num* num1, int pow);
 
-float abstract_value(complex_num * num);
+double abstract_value(complex_num * num);
